refactor: simplify RP authorization transport
