+++ docs/WIPs/wip-101.md
@@ -91,7 +91,7 @@ This specification introduces both concerns of "Authorization" and "Validation" 
 
 The version of the [`ProofRequest`](https://docs.rs/world-id-primitives/latest/world_id_primitives/request/struct.ProofRequest.html) is defined in the [`RequestVersion`](https://docs.rs/world-id-primitives/latest/world_id_primitives/request/enum.RequestVersion.html) enum and is included in `RpRequest`.
 
-The complete EIP-712 authorization and private-details commitment are specified in [RP request authorization](../world-id-4-specs/rp-request-authorization.md). A contract receives only the public intent and `detailsHash` by default. If it needs to validate private request details, `data` MAY carry those details and the salt in an RP-defined encoding; the contract SHOULD recompute and compare `detailsHash` before relying on them.
+The complete EIP-712 authorization and private-details commitment are specified in [RP request authorization](../world-id-4-specs/rp-request-authorization.md). A contract receives only the public intent and `detailsHash` by default, so it MUST treat the hash as an opaque commitment rather than evidence of any particular private fields. Passing private details is optional, but if the contract's authorization decision depends on them, `data` MUST carry those details and the salt in an RP-defined encoding and the contract MUST recompute and compare `detailsHash` before relying on them. This comparison validates the commitment supplied to the contract; it does not make `detailsHash` a public input to the query proof or final World ID proof.
 
 **Invalidating Signer Cache.**
