+++ docs/world-id-4-specs/rp-request-authorization.md
@@ -73,6 +73,8 @@ struct RpRequestDetails {
 
 The RP MUST supply a fresh, unpredictable, non-zero 32-byte `request_salt`. It is sent to the Authenticator with the proof request but is not forwarded to OPRF nodes. Salting prevents a node from guessing a low-entropy private request body from `detailsHash`.
 
+The Authenticator recomputes `detailsHash` from the exact `ProofRequest` it received, while OPRF nodes validate the RP authorization over that hash without receiving or interpreting its preimage. With an honest Authenticator, a party between the RP and Authenticator therefore cannot substitute private request fields while retaining a valid authorization: the recomputed hash would no longer match the signed hash. OPRF nodes intentionally treat the commitment as opaque and cannot perform this comparison themselves.
+
 The nested hashes commit to:
 
 - the exact request ID;
@@ -83,6 +85,8 @@ The nested hashes commit to:
 Optional fields are normalized to their protocol meaning before hashing: an absent `genesisIssuedAtMin` is zero, and an absent `expiresAtMin` is the request's `createdAt`. Explicit presence tags distinguish values where absence has different semantics, such as the signal and action.
 
 The request signature, query proof, Merkle root, and values generated by the Authenticator for a new session are not part of the private details commitment. They are either the authorization itself, independently verified by the nodes, or unavailable when the RP signs. Their permitted behavior is instead constrained by the signed public operation matrix.
+
+The commitment does not protect against a malicious or compromised Authenticator because `detailsHash` is not a public input to the query proof or final World ID proof. Such an Authenticator could forward a valid authorization while generating a proof from different private semantics. For contract-backed RPs, WIP-101 provides contract-side validation of private fields only when the contract receives those fields and the salt through `data` and recomputes `detailsHash`; a contract that does not recompute it is authorizing an opaque commitment.
 
 For an existing session, the Authenticator discloses a request-specific opening only to a node call that rederives the seed. The node verifies:
